LibAudio+LibDSP: Switch samples to 32-bit float instead of 64-bit float
This has been overkill from the start, and it has been bugging me for a
long time. With this change, we're probably a bit slower on most
platforms but save huge amounts of space with all in-memory sample
datastructures.

+++ Userland/Applications/SoundPlayer/AlbumCoverVisualizationWidget.h
@@ -18,7 +18,7 @@ public:
     void start_new_file(StringView) override;
 
 private:
-    void render(GUI::PaintEvent&, FixedArray<double> const&) override { }
+    void render(GUI::PaintEvent&, FixedArray<float> const&) override { }
     void paint_event(GUI::PaintEvent&) override;
     AlbumCoverVisualizationWidget() = default;
     ErrorOr<NonnullRefPtr<Gfx::Bitmap>> get_album_cover(StringView const filename);
