+++ Userland/Applications/SoundPlayer/BarsVisualizationWidget.h
@@ -22,7 +22,7 @@ public:
 private:
     BarsVisualizationWidget();
 
-    void render(GUI::PaintEvent&, FixedArray<double> const&) override;
+    void render(GUI::PaintEvent&, FixedArray<float> const&) override;
     void context_menu_event(GUI::ContextMenuEvent& event) override;
 
     static constexpr size_t fft_size = 512;
@@ -30,9 +30,9 @@ private:
     // Things become weird near the Nyquist limit. Just don't use that FFT data.
     static constexpr size_t cutoff = fft_size - 32;
 
-    Array<Complex<double>, fft_size> m_fft_samples {};
-    Array<double, fft_size> m_fft_window {};
-    Array<double, fft_size / 2> m_previous_samples {};
+    Array<Complex<float>, fft_size> m_fft_samples {};
+    Array<float, fft_size> m_fft_window {};
+    Array<float, fft_size / 2> m_previous_samples {};
     Array<int, bar_count> m_gfx_falling_bars {};
     bool m_is_using_last;
     bool m_adjust_frequencies;
